chore: Rename `shared` to `util`
