+++ internal/util/git/git_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/numaproj-labs/numaplane/internal/shared/kubernetes"
+	"github.com/numaproj-labs/numaplane/internal/util/kubernetes"
 )
 
 func TestCheckGitURL(t *testing.T) {
