+++ internal/util/kubernetes/testdata/svc-with-invalid-data.yaml
