+++ internal/util/kubernetes/testdata/svc.yaml
